requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
pypdf==4.1.0
pikepdf==8.13.0
chromadb==0.4.24
unstructured==0.12.4
unstructured-inference==0.7.24
pdf2image==1.17.0
pillow==10.2.0
pillow-heif==0.15.0
streamlit==1.32.0
pdfminer.six==20231228
python-magic-bin==0.4.14; platform_system == 'Windows'
python-magic==0.4.27; platform_system != 'Windows'
nltk==3.8.1
markdown==3.5.2
# poppler-utils==0.1.0  # For Linux
# tesseract-ocr  # For Linux
# hnswlib==0.7.0
sentence-transformers==2.5.1
opencv-python==4.9.0.80